Rename: #341 - 불필요한 코드 삭제 및 이동
