+++ src/main/java/com/teamddd/duckmap/advice/ServiceAdvice.java
@@ -7,9 +7,7 @@ import org.springframework.web.bind.annotation.RestControllerAdvice;
 
 import com.teamddd.duckmap.common.ExceptionCodeMessage;
 import com.teamddd.duckmap.dto.ErrorResult;
-import com.teamddd.duckmap.exception.NonExistentEventLikeException;
 import com.teamddd.duckmap.exception.NonExistentFileException;
-import com.teamddd.duckmap.exception.NonExistentMemberException;
 import com.teamddd.duckmap.exception.ServiceException;
 
 import lombok.extern.slf4j.Slf4j;
@@ -36,21 +34,4 @@ public class ServiceAdvice {
 			.build();
 	}
 
-	@ResponseStatus(HttpStatus.BAD_REQUEST)
-	@ExceptionHandler
-	public ErrorResult nonExistentMemberException(NonExistentMemberException ex) {
-		return ErrorResult.builder()
-			.code(ExceptionCodeMessage.NON_EXISTENT_MEMBER_EXCEPTION.code())
-			.message(ex.getMessage())
-			.build();
-	}
-
-	@ResponseStatus(HttpStatus.BAD_REQUEST)
-	@ExceptionHandler
-	public ErrorResult nonExistentEventLikeMemberException(NonExistentEventLikeException ex) {
-		return ErrorResult.builder()
-			.code(ExceptionCodeMessage.NON_EXISTENT_EVENT_LIKE_EXCEPTION.code())
-			.message(ex.getMessage())
-			.build();
-	}
 }
